requirements.txt
highrise-bot-sdk
websockets>=10.4
python-dotenv>=0.19.0
